refactor: move fba format/file functions to sep file
