+++ src/extension/fbaNBRQRenderer.ts
@@ -13,7 +13,8 @@ import * as jp from 'jsonpath'
 // import jju from 'jju'
 const jju = require('jju')
 import { RawNotebookCell } from './fbaNotebookProvider'
-import { DocData, FBAEditorProvider, FBBadge } from './fbaEditor'
+import { DocData, FBAEditorProvider } from './fbaEditor'
+import { FBBadge } from './fbaFormat'
 import { arrayEquals, getMemberParent, MemberPath } from './util'
 import * as uv0 from 'dlt-logs-utils'
 import { NotebookCellOutput } from 'vscode'
